#3653: Made Brent solver into a separate block solver that can be enabled by experiment_mode = 4 (jmi_block_solver_experimental_Brent).  The code compiles and is complete with error checking. Not tested.
git-svn-id: https://svn.jmodelica.org@6333 7ae51c65-13d5-4921-9d80-6511ddd1419c

+++ trunk/RuntimeLibrary/src/jmi/jmi_me.c
@@ -194,13 +194,13 @@ int jmi_set_real(jmi_t* jmi, const jmi_value_reference vr[], size_t nvr,
         /* Get index in z vector from value reference. */
         index = get_index_from_value_ref(vr[i]);
 
-        if (index >= jmi->offs_real_pd && index < jmi->offs_integer_ci) {
+        if (index >= (size_t)(jmi->offs_real_pd) && index < (size_t)(jmi->offs_integer_ci)) {
             jmi_log_node(jmi->log, logError, "CannotSetVariable",
                          "Cannot set Real dependent parameter <variable: #r%d#>", vr[i]);
             return -1;
         }
 
-        if (index >= jmi->offs_real_ci && index < jmi->offs_real_pi) {
+        if (index >= (size_t)jmi->offs_real_ci && index < (size_t)jmi->offs_real_pi) {
             jmi_log_node(jmi->log, logError, "CannotSetVariable",
                          "Cannot set Real constant <variable: #r%d#>", vr[i]);
             return -1;
